Reorganizes the staking folder structure.
