+++ lib/screens/home/explorer/staking_flow/staking_confirm/confirm_claim_rewards_screen.dart
@@ -2,16 +2,15 @@ import 'package:provenance_wallet/common/pw_design.dart';
 import 'package:provenance_wallet/common/widgets/modal_loading.dart';
 import 'package:provenance_wallet/common/widgets/pw_list_divider.dart';
 import 'package:provenance_wallet/dialogs/error_dialog.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_confirm/staking_confirm_base.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_delegation/staking_delegation_bloc.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_confirm/staking_confirm_base.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_delegation/staking_delegation_bloc.dart';
 import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_flow.dart';
 import 'package:provenance_wallet/screens/home/transactions/details_item.dart';
-import 'package:provenance_wallet/util/extensions/num_extensions.dart';
 import 'package:provenance_wallet/util/get.dart';
 import 'package:provenance_wallet/util/strings.dart';
 
-class ConfirmDelegateScreen extends StatelessWidget {
-  const ConfirmDelegateScreen({Key? key, required this.navigator})
+class ConfirmClaimRewardsScreen extends StatelessWidget {
+  const ConfirmClaimRewardsScreen({Key? key, required this.navigator})
       : super(key: key);
 
   final StakingFlowNavigator navigator;
@@ -28,16 +27,13 @@ class ConfirmDelegateScreen extends StatelessWidget {
         if (details == null) {
           return Container();
         }
+
         return StakingConfirmBase(
           appBarTitle: details.selectedDelegationType.dropDownTitle,
           onDataClick: () {
             final data = '''{
   "delegatorAddress": "${details.account.address}",
   "validatorAddress": "${details.validator.operatorAddress}",
-  "amount": {
-    "denom": "nhash",
-    "amount": "${details.hashDelegated.nhashFromHash()}"
-  }
 }''';
             navigator.showTransactionData(data);
           },
@@ -45,9 +41,10 @@ class ConfirmDelegateScreen extends StatelessWidget {
             ModalLoadingRoute.showLoading('', context);
             // Give the loading modal time to display
             await Future.delayed(Duration(milliseconds: 500));
-            await _sendTransaction(bloc, details, gasAdjustment, context);
+            await _sendTransaction(
+                gasAdjustment, details.selectedDelegationType, context);
           },
-          signButtonTitle: details.selectedDelegationType.dropDownTitle,
+          signButtonTitle: Strings.stakingConfirmRewardClaim,
           children: [
             DetailsItem(
               title: Strings.stakingConfirmDelegatorAddress,
@@ -76,39 +73,6 @@ class ConfirmDelegateScreen extends StatelessWidget {
                 ),
               ),
             ),
-            PwListDivider(
-              indent: Spacing.largeX3,
-            ),
-            DetailsItem(
-              title: Strings.stakingConfirmDenom,
-              endChild: Flexible(
-                child: PwText(
-                  details.asset?.denom ?? Strings.stakingConfirmHash,
-                  overflow: TextOverflow.fade,
-                  softWrap: false,
-                  color: PwColor.neutralNeutral,
-                  style: PwTextStyle.body,
-                ),
-              ),
-            ),
-            PwListDivider(
-              indent: Spacing.largeX3,
-            ),
-            DetailsItem(
-              title: Strings.stakingConfirmAmount,
-              endChild: Flexible(
-                child: PwText(
-                  details.hashDelegated.nhashFromHash(),
-                  overflow: TextOverflow.fade,
-                  softWrap: false,
-                  color: PwColor.neutralNeutral,
-                  style: PwTextStyle.body,
-                ),
-              ),
-            ),
-            PwListDivider(
-              indent: Spacing.largeX3,
-            ),
           ],
         );
       },
@@ -116,40 +80,24 @@ class ConfirmDelegateScreen extends StatelessWidget {
   }
 
   Future<void> _sendTransaction(
-    StakingDelegationBloc bloc,
-    StakingDelegationDetails details,
     double? gasAdjustment,
+    SelectedDelegationType selected,
     BuildContext context,
   ) async {
-    final selected = details.selectedDelegationType;
-    if (SelectedDelegationType.delegate == selected) {
-      try {
-        await bloc.doDelegate(gasAdjustment);
-        ModalLoadingRoute.dismiss(context);
-        navigator.showTransactionSuccess(selected);
-      } catch (err) {
-        await _showErrorModal(err, context);
-      }
-    } else {
-      try {
-        await bloc.doUndelegate(gasAdjustment);
-        ModalLoadingRoute.dismiss(context);
-        navigator.showTransactionSuccess(selected);
-      } catch (err) {
-        await _showErrorModal(err, context);
-      }
+    try {
+      await (get<StakingDelegationBloc>()).claimRewards(gasAdjustment);
+      ModalLoadingRoute.dismiss(context);
+      navigator.showTransactionSuccess(selected);
+    } catch (err) {
+      ModalLoadingRoute.dismiss(context);
+      showDialog(
+        context: context,
+        builder: (context) {
+          return ErrorDialog(
+            error: err.toString(),
+          );
+        },
+      );
     }
-  }
-
-  Future<void> _showErrorModal(Object error, BuildContext context) async {
-    ModalLoadingRoute.dismiss(context);
-    await showDialog(
-      context: context,
-      builder: (context) {
-        return ErrorDialog(
-          error: error.toString(),
-        );
-      },
-    );
   }
 }
