+++ lib/screens/home/explorer/staking_flow/staking_confirm/confirm_delegate_screen.dart
@@ -2,28 +2,27 @@ import 'package:provenance_wallet/common/pw_design.dart';
 import 'package:provenance_wallet/common/widgets/modal_loading.dart';
 import 'package:provenance_wallet/common/widgets/pw_list_divider.dart';
 import 'package:provenance_wallet/dialogs/error_dialog.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_confirm/staking_confirm_base.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_delegation/staking_delegation_bloc.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_confirm/staking_confirm_base.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_delegation/staking_delegation_bloc.dart';
 import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_flow.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_redelegation/staking_redelegation_bloc.dart';
 import 'package:provenance_wallet/screens/home/transactions/details_item.dart';
 import 'package:provenance_wallet/util/extensions/num_extensions.dart';
 import 'package:provenance_wallet/util/get.dart';
 import 'package:provenance_wallet/util/strings.dart';
 
-class ConfirmRedelegateScreen extends StatelessWidget {
-  const ConfirmRedelegateScreen({Key? key, required this.navigator})
+class ConfirmDelegateScreen extends StatelessWidget {
+  const ConfirmDelegateScreen({Key? key, required this.navigator})
       : super(key: key);
 
   final StakingFlowNavigator navigator;
 
   @override
   Widget build(BuildContext context) {
-    final bloc = get<StakingRedelegationBloc>();
+    final bloc = get<StakingDelegationBloc>();
 
-    return StreamBuilder<StakingRedelegationDetails>(
-      initialData: bloc.stakingRedelegationDetails.value,
-      stream: bloc.stakingRedelegationDetails,
+    return StreamBuilder<StakingDelegationDetails>(
+      initialData: bloc.stakingDelegationDetails.value,
+      stream: bloc.stakingDelegationDetails,
       builder: (context, snapshot) {
         final details = snapshot.data;
         if (details == null) {
@@ -32,29 +31,21 @@ class ConfirmRedelegateScreen extends StatelessWidget {
         return StakingConfirmBase(
           appBarTitle: details.selectedDelegationType.dropDownTitle,
           onDataClick: () {
-            final data = '''
-{
+            final data = '''{
   "delegatorAddress": "${details.account.address}",
-  "validatorSrcAddress": "${details.delegation.sourceAddress}",
-  "validatorDstAddress": "${details.toRedelegate?.addressId}",
+  "validatorAddress": "${details.validator.operatorAddress}",
   "amount": {
     "denom": "nhash",
-    "amount": "${details.hashRedelegated.nhashFromHash()}"
+    "amount": "${details.hashDelegated.nhashFromHash()}"
   }
-}
-''';
+}''';
             navigator.showTransactionData(data);
           },
           onTransactionSign: (gasAdjustment) async {
             ModalLoadingRoute.showLoading('', context);
             // Give the loading modal time to display
             await Future.delayed(Duration(milliseconds: 500));
-            await _sendTransaction(
-              bloc,
-              details.selectedDelegationType,
-              gasAdjustment,
-              context,
-            );
+            await _sendTransaction(bloc, details, gasAdjustment, context);
           },
           signButtonTitle: details.selectedDelegationType.dropDownTitle,
           children: [
@@ -74,25 +65,10 @@ class ConfirmRedelegateScreen extends StatelessWidget {
               indent: Spacing.largeX3,
             ),
             DetailsItem(
-              title: Strings.stakingConfirmValidatorSource,
+              title: Strings.stakingConfirmValidatorAddress,
               endChild: Flexible(
                 child: PwText(
-                  details.delegation.sourceAddress.abbreviateAddress(),
-                  overflow: TextOverflow.fade,
-                  softWrap: false,
-                  color: PwColor.neutralNeutral,
-                  style: PwTextStyle.body,
-                ),
-              ),
-            ),
-            PwListDivider(
-              indent: Spacing.largeX3,
-            ),
-            DetailsItem(
-              title: Strings.stakingConfirmValidatorDestination,
-              endChild: Flexible(
-                child: PwText(
-                  details.toRedelegate?.addressId.abbreviateAddress() ?? "",
+                  details.validator.operatorAddress.abbreviateAddress(),
                   overflow: TextOverflow.fade,
                   softWrap: false,
                   color: PwColor.neutralNeutral,
@@ -107,7 +83,7 @@ class ConfirmRedelegateScreen extends StatelessWidget {
               title: Strings.stakingConfirmDenom,
               endChild: Flexible(
                 child: PwText(
-                  Strings.stakingConfirmHash,
+                  details.asset?.denom ?? Strings.stakingConfirmHash,
                   overflow: TextOverflow.fade,
                   softWrap: false,
                   color: PwColor.neutralNeutral,
@@ -122,7 +98,7 @@ class ConfirmRedelegateScreen extends StatelessWidget {
               title: Strings.stakingConfirmAmount,
               endChild: Flexible(
                 child: PwText(
-                  details.hashRedelegated.nhashFromHash(),
+                  details.hashDelegated.nhashFromHash(),
                   overflow: TextOverflow.fade,
                   softWrap: false,
                   color: PwColor.neutralNeutral,
@@ -140,25 +116,40 @@ class ConfirmRedelegateScreen extends StatelessWidget {
   }
 
   Future<void> _sendTransaction(
-    StakingRedelegationBloc bloc,
-    SelectedDelegationType selected,
+    StakingDelegationBloc bloc,
+    StakingDelegationDetails details,
     double? gasAdjustment,
     BuildContext context,
   ) async {
-    try {
-      await (get<StakingRedelegationBloc>()).doRedelegate(gasAdjustment);
-      ModalLoadingRoute.dismiss(context);
-      navigator.showTransactionSuccess(selected);
-    } catch (err) {
-      ModalLoadingRoute.dismiss(context);
-      showDialog(
-        context: context,
-        builder: (context) {
-          return ErrorDialog(
-            error: err.toString(),
-          );
-        },
-      );
+    final selected = details.selectedDelegationType;
+    if (SelectedDelegationType.delegate == selected) {
+      try {
+        await bloc.doDelegate(gasAdjustment);
+        ModalLoadingRoute.dismiss(context);
+        navigator.showTransactionSuccess(selected);
+      } catch (err) {
+        await _showErrorModal(err, context);
+      }
+    } else {
+      try {
+        await bloc.doUndelegate(gasAdjustment);
+        ModalLoadingRoute.dismiss(context);
+        navigator.showTransactionSuccess(selected);
+      } catch (err) {
+        await _showErrorModal(err, context);
+      }
     }
+  }
+
+  Future<void> _showErrorModal(Object error, BuildContext context) async {
+    ModalLoadingRoute.dismiss(context);
+    await showDialog(
+      context: context,
+      builder: (context) {
+        return ErrorDialog(
+          error: error.toString(),
+        );
+      },
+    );
   }
 }
