+++ lib/screens/home/explorer/staking_flow/staking_confirm/confirm_redelegate_screen.dart
@@ -1,0 +1,164 @@
+import 'package:provenance_wallet/common/pw_design.dart';
+import 'package:provenance_wallet/common/widgets/modal_loading.dart';
+import 'package:provenance_wallet/common/widgets/pw_list_divider.dart';
+import 'package:provenance_wallet/dialogs/error_dialog.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_confirm/staking_confirm_base.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_delegation/staking_delegation_bloc.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_flow.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_redelegation/staking_redelegation_bloc.dart';
+import 'package:provenance_wallet/screens/home/transactions/details_item.dart';
+import 'package:provenance_wallet/util/extensions/num_extensions.dart';
+import 'package:provenance_wallet/util/get.dart';
+import 'package:provenance_wallet/util/strings.dart';
+
+class ConfirmRedelegateScreen extends StatelessWidget {
+  const ConfirmRedelegateScreen({Key? key, required this.navigator})
+      : super(key: key);
+
+  final StakingFlowNavigator navigator;
+
+  @override
+  Widget build(BuildContext context) {
+    final bloc = get<StakingRedelegationBloc>();
+
+    return StreamBuilder<StakingRedelegationDetails>(
+      initialData: bloc.stakingRedelegationDetails.value,
+      stream: bloc.stakingRedelegationDetails,
+      builder: (context, snapshot) {
+        final details = snapshot.data;
+        if (details == null) {
+          return Container();
+        }
+        return StakingConfirmBase(
+          appBarTitle: details.selectedDelegationType.dropDownTitle,
+          onDataClick: () {
+            final data = '''
+{
+  "delegatorAddress": "${details.account.address}",
+  "validatorSrcAddress": "${details.delegation.sourceAddress}",
+  "validatorDstAddress": "${details.toRedelegate?.addressId}",
+  "amount": {
+    "denom": "nhash",
+    "amount": "${details.hashRedelegated.nhashFromHash()}"
+  }
+}
+''';
+            navigator.showTransactionData(data);
+          },
+          onTransactionSign: (gasAdjustment) async {
+            ModalLoadingRoute.showLoading('', context);
+            // Give the loading modal time to display
+            await Future.delayed(Duration(milliseconds: 500));
+            await _sendTransaction(
+              bloc,
+              details.selectedDelegationType,
+              gasAdjustment,
+              context,
+            );
+          },
+          signButtonTitle: details.selectedDelegationType.dropDownTitle,
+          children: [
+            DetailsItem(
+              title: Strings.stakingConfirmDelegatorAddress,
+              endChild: Flexible(
+                child: PwText(
+                  details.account.address.abbreviateAddress(),
+                  overflow: TextOverflow.fade,
+                  softWrap: false,
+                  color: PwColor.neutralNeutral,
+                  style: PwTextStyle.body,
+                ),
+              ),
+            ),
+            PwListDivider(
+              indent: Spacing.largeX3,
+            ),
+            DetailsItem(
+              title: Strings.stakingConfirmValidatorSource,
+              endChild: Flexible(
+                child: PwText(
+                  details.delegation.sourceAddress.abbreviateAddress(),
+                  overflow: TextOverflow.fade,
+                  softWrap: false,
+                  color: PwColor.neutralNeutral,
+                  style: PwTextStyle.body,
+                ),
+              ),
+            ),
+            PwListDivider(
+              indent: Spacing.largeX3,
+            ),
+            DetailsItem(
+              title: Strings.stakingConfirmValidatorDestination,
+              endChild: Flexible(
+                child: PwText(
+                  details.toRedelegate?.addressId.abbreviateAddress() ?? "",
+                  overflow: TextOverflow.fade,
+                  softWrap: false,
+                  color: PwColor.neutralNeutral,
+                  style: PwTextStyle.body,
+                ),
+              ),
+            ),
+            PwListDivider(
+              indent: Spacing.largeX3,
+            ),
+            DetailsItem(
+              title: Strings.stakingConfirmDenom,
+              endChild: Flexible(
+                child: PwText(
+                  Strings.stakingConfirmHash,
+                  overflow: TextOverflow.fade,
+                  softWrap: false,
+                  color: PwColor.neutralNeutral,
+                  style: PwTextStyle.body,
+                ),
+              ),
+            ),
+            PwListDivider(
+              indent: Spacing.largeX3,
+            ),
+            DetailsItem(
+              title: Strings.stakingConfirmAmount,
+              endChild: Flexible(
+                child: PwText(
+                  details.hashRedelegated.nhashFromHash(),
+                  overflow: TextOverflow.fade,
+                  softWrap: false,
+                  color: PwColor.neutralNeutral,
+                  style: PwTextStyle.body,
+                ),
+              ),
+            ),
+            PwListDivider(
+              indent: Spacing.largeX3,
+            ),
+          ],
+        );
+      },
+    );
+  }
+
+  Future<void> _sendTransaction(
+    StakingRedelegationBloc bloc,
+    SelectedDelegationType selected,
+    double? gasAdjustment,
+    BuildContext context,
+  ) async {
+    try {
+      await (get<StakingRedelegationBloc>()).doRedelegate(gasAdjustment);
+      ModalLoadingRoute.dismiss(context);
+      navigator.showTransactionSuccess(selected);
+    } catch (err) {
+      ModalLoadingRoute.dismiss(context);
+      showDialog(
+        context: context,
+        builder: (context) {
+          return ErrorDialog(
+            error: err.toString(),
+          );
+        },
+      );
+    }
+  }
+}
