+++ lib/screens/home/explorer/staking_flow/staking_confirm/staking_confirm_base.dart
