+++ lib/screens/home/explorer/staking_flow/staking_details/details_item_copy.dart
