+++ lib/screens/home/explorer/staking_flow/staking_details/staking_details_bloc.dart
