+++ lib/screens/home/explorer/staking_flow/staking_details/staking_details_screen.dart
@@ -2,9 +2,9 @@ import 'package:provenance_wallet/common/pw_design.dart';
 import 'package:provenance_wallet/common/widgets/button.dart';
 import 'package:provenance_wallet/common/widgets/pw_app_bar.dart';
 import 'package:provenance_wallet/common/widgets/pw_list_divider.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_details/details_item_copy.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_details/staking_details_bloc.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_details/staking_management.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_details/details_item_copy.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_details/staking_details_bloc.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_details/staking_management.dart';
 import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_flow.dart';
 import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_flow_bloc.dart';
 import 'package:provenance_wallet/screens/home/transactions/details_item.dart';
