+++ lib/screens/home/explorer/staking_flow/staking_flow.dart
@@ -1,15 +1,15 @@
 import 'package:provenance_wallet/common/flow_base.dart';
 import 'package:provenance_wallet/common/pw_design.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_confirm/confirm_claim_rewards_screen.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_confirm/confirm_delegate_screen.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_confirm/confirm_redelegate_screen.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_confirm/staking_transaction_data_screen.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_delegation/staking_delegation_bloc.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_delegation/staking_delegation_screen.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_delegation/staking_undelegation_screen.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_details/staking_details_screen.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_redelegation/staking_redelegation_screen.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_success/staking_success_screen.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_confirm/confirm_claim_rewards_screen.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_confirm/confirm_delegate_screen.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_confirm/confirm_redelegate_screen.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_confirm/staking_transaction_data_screen.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_delegation/staking_delegation_bloc.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_delegation/staking_delegation_screen.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_delegation/staking_undelegation_screen.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_details/staking_details_screen.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_redelegation/staking_redelegation_screen.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_success/staking_success_screen.dart';
 import 'package:provenance_wallet/services/models/account.dart';
 import 'package:provenance_wallet/services/models/commission.dart';
 import 'package:provenance_wallet/services/models/delegation.dart';
