+++ lib/screens/home/explorer/staking_flow/staking_redelegation/redelegation_list.dart
@@ -1,7 +1,7 @@
 import 'package:provenance_wallet/common/pw_design.dart';
 import 'package:provenance_wallet/common/widgets/pw_list_divider.dart';
 import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_flow_bloc.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_redelegation/redelegation_list_item.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_redelegation/redelegation_list_item.dart';
 import 'package:provenance_wallet/services/models/detailed_validator.dart';
 import 'package:provenance_wallet/util/get.dart';
