+++ lib/screens/home/explorer/staking_flow/staking_redelegation/staking_redelegation_screen.dart
@@ -3,9 +3,9 @@ import 'package:provenance_wallet/common/widgets/pw_dropdown.dart';
 import 'package:provenance_wallet/common/widgets/pw_list_divider.dart';
 import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_flow.dart';
 import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_flow_bloc.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_redelegation/redelegation_list.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_redelegation/staking_redelegation_bloc.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_redelegation/staking_redelegation_list.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_redelegation/redelegation_list.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_redelegation/staking_redelegation_bloc.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_flow/staking_redelegation/staking_redelegation_list.dart';
 import 'package:provenance_wallet/screens/home/transactions/details_item.dart';
 import 'package:provenance_wallet/services/models/account.dart';
 import 'package:provenance_wallet/services/models/delegation.dart';
